+++ web/io/request/training.go
@@ -9,13 +9,13 @@ type List struct {
 
 type ListProblem struct {
 	LID int64 `json:"LID"`
-	PID int   `json:"PID"`
+	PID int64 `json:"PID"`
 }
 
 type ListAll struct {
 	LID       int64  `json:"LID"`
 	UID       string `json:"UID"`
-	PID       int    `json:"PID"`
+	PID       int64  `json:"PID"`
 	Title     string `json:"Title"`
 	StartTime int64  `json:"StartTime"`
 }
